Move WidgetChoices to widget file
