+++ app/grandchallenge/cases/utils.py
@@ -3,7 +3,6 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 
 from django.conf import settings
-from django.db.models import TextChoices
 from panimg.image_builders.metaio_utils import load_sitk_image
 
 
@@ -53,8 +52,3 @@ def get_sitk_image(*, image):
             raise
 
     return sitk_image
-
-
-class WidgetChoices(TextChoices):
-    IMAGE_SEARCH = "IMAGE_SEARCH"
-    IMAGE_UPLOAD = "IMAGE_UPLOAD"
