+++ app/tests/cases_tests/test_widget.py
@@ -2,8 +2,7 @@ import pytest
 from django.core.exceptions import ValidationError
 from guardian.shortcuts import assign_perm
 
-from grandchallenge.cases.utils import WidgetChoices
-from grandchallenge.cases.widgets import FlexibleImageField
+from grandchallenge.cases.widgets import FlexibleImageField, WidgetChoices
 from grandchallenge.components.models import ComponentInterface
 from grandchallenge.core.guardian import get_objects_for_user
 from grandchallenge.uploads.models import UserUpload
